Show event list next to calendar
