+++ src/App.js
@@ -1,12 +1,21 @@
 import './App.css';
 import Calendar from './Calendar';
 import EventManager from './EventManager';
-import { Container } from '@mui/material';
-import { useState } from 'react';
+import EventList from './EventList';
+import { Container, Box } from '@mui/material';
+import { useState, useEffect } from 'react';
 
 function App() {
   const [dialogOpen, setDialogOpen] = useState(false);
   const [selectedDate, setSelectedDate] = useState(null);
+  const [events, setEvents] = useState(() => {
+    const stored = localStorage.getItem('events');
+    return stored ? JSON.parse(stored) : [];
+  });
+
+  useEffect(() => {
+    localStorage.setItem('events', JSON.stringify(events));
+  }, [events]);
 
   const handleDateClick = (date) => {
     setSelectedDate(date);
@@ -19,11 +28,16 @@ function App() {
 
   return (
     <Container className="App">
-      <Calendar onDateClick={handleDateClick} />
+      <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'flex-start', gap: 4 }}>
+        <Calendar onDateClick={handleDateClick} />
+        <EventList events={events} />
+      </Box>
       <EventManager
         open={dialogOpen}
         onClose={handleClose}
         defaultDate={selectedDate}
+        events={events}
+        setEvents={setEvents}
       />
     </Container>
   );
